feat: manifest-only
