--- airbyte-integrations/connectors/source-google-sheets/.coveragerc
@@ -1,3 +1,0 @@
-[run]
-omit =
-    source_google_sheets/run.py
