--- airbyte-integrations/connectors/source-google-sheets/main.py
@@ -1,9 +1,0 @@
-#
-# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
-#
-
-from source_google_sheets.run import run
-
-
-if __name__ == "__main__":
-    run()
