--- airbyte-integrations/connectors/source-google-sheets/pyproject.toml
@@ -1,35 +1,0 @@
-[build-system]
-requires = [ "poetry-core>=1.0.0",]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry]
-version = "0.11.0"
-name = "source-google-sheets"
-description = "Source implementation for Google Sheets."
-authors = [ "Airbyte <contact@airbyte.io>",]
-license = "Elv2"
-readme = "README.md"
-documentation = "https://docs.airbyte.com/integrations/sources/google-sheets"
-homepage = "https://airbyte.com"
-repository = "https://github.com/airbytehq/airbyte"
-[[tool.poetry.packages]]
-include = "source_google_sheets"
-
-[tool.poetry.dependencies]
-python = ">=3.10,<3.13"
-airbyte-cdk = "^6"
-
-[tool.poetry.scripts]
-source-google-sheets = "source_google_sheets.run:run"
-
-[tool.poetry.group.dev.dependencies]
-requests-mock = "^1.9.3"
-pytest-mock = "^3.6.1"
-pytest = "^8.0.0"
-
-[tool.poe]
-include = [
-    # Shared tasks definition file(s) can be imported here.
-    # Run `poe` or `poe --help` to see the list of available tasks.
-    "${POE_GIT_DIR}/poe-tasks/poetry-connector-tasks.toml",
-]
