--- airbyte-integrations/connectors/source-google-sheets/setup.cfg
@@ -1,2 +1,0 @@
-[aliases]
-test='pytest'
