--- airbyte-integrations/connectors/source-google-sheets/source_google_sheets/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
-#
-
-from .source import SourceGoogleSheets
-
-__all__ = ["SourceGoogleSheets"]
