--- airbyte-integrations/connectors/source-google-sheets/source_google_sheets/components/__init__.py
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
-#
-
-from source_google_sheets.components.extractors import DpathSchemaMatchingExtractor, DpathSchemaExtractor
-from source_google_sheets.components.partition_routers import RangePartitionRouter
-
-__all__ = ["DpathSchemaMatchingExtractor", "RangePartitionRouter", "DpathSchemaExtractor"]
