--- airbyte-integrations/connectors/source-google-sheets/source_google_sheets/models/__init__.py
@@ -1,2 +1,0 @@
-from .spreadsheet import *
-from .spreadsheet_values import *
